Only respond to Give Physical Address with Report Physical Address. bugzid: 592
This fixes several reports of the TV improperly switching to the libcec
device after it has previously switched to another source. Fix confirmed
with Panasonic and also reported to be an issue with Samsung and
Philips.

The previous extra processing (SetActiveSource and ImageViewOn) was put
in place for Samsung and may need revisiting.