LibCECTray updates and fixes
Added ability to send modifiers by nesting commands in KeyInput.cs

Made it so that if you click the first column in the WMC grid view. It'll send the command to the open app.

Fixed bug with libcec adding the keyup for button commands and making sure they are still executed.