Removes .egplugin binary file
New method of creating the .egplugin file through create_installer.cmd. Power shell is used to do this since windows zip compression is not accessable from a command line. 

I have also added file cleanup to ensure any fils that have been copied to create the .egplugin file were removed to leave the system in the state it was before the installer was created.

I also added checking to make sure the .egplugin file was created. and if it was not then to exit the build process.