Fix broken Python version check and failure to build on cmake < 3.7
Ubuntu 16.04 and derivative distributions ship cmake 3.5, which does
not support GREATER_EQUAL.  Tested fix on Ubuntu MATE 16.04 for
Raspberry Pi.

Also, the existing logic to check for Python 2.7+ does not work.
This patch corrects that, and I was able to 'import cec' in
Python 3 on Ubuntu MATE 16.04 for Raspberry Pi without error.