fixed: fix build if hostname does not support '-f' argument
Part of the build information compiled into the library is the host on
which it was built. This information is generated by executing the
command `hostname -f`, but the '-f' argument may not be available on all
platforms (e.g. coreutils v8.30 does not have it). As a result, the
command will spit an error message that spans multiple lines, break the
format of the define in "src/libcec/env.h" and as a result the error
will fail due to invalid syntax.

Fix the issue by falling back to just `hostname` if `hostname -f`
returns an error.