Linux Adapter: support change in logical address
when multiple devices of the same type (e.g. recoder) are connected to a TV the second
device needs a different logical address. Without that change each
device gets the same logical address in the libcec library, but Linux
kernel assigns a different logical address. This patch de-allocats the
initial logical device being used and re-syncs to the logical address
being used by the kernel