Add an udev rule making the CEC adapter part of the input group
Without this, the HDMI CEC USB adapter appears as /dev/ttyACM0 owned by
the dialout group with mode 660. This means that any user that wants to
use this device needs to be part of the dialout group. This actually has
security concerns as all serial devices are owned by that group as well
and it would give the user too much control of them, meaning any rogue
process can cause harm without needing root access.

Since the CEC adapter is actually used as an input device, this udev
rule makes the device be part of the input group instead. This makes
sense as it _is_ an input device and any rogue process shouldn't be able
to do to much harm with that group. On most distributions the standard
user is part of that group by default anyway